🚚 move nav components under ChatsListItem and User components accordingly
